Log more activity.
This generically logs all mousemove, mousedown, mouseup, click, keydown,
and keyup events, capturing most of the user activity.  Events are
logged with where they occur.

Experiments are logged distinctly.

We will probably want to hook up other semantic logging -- you can tell
if the user clicks a control currently, but it might be better to log a
control-specific message.  Similarly, you can reconstruct if an
annotation is drawn, but it might be better to log that the annotation
state and count has changed.