Refactor docker-compose.
This allows running girder_worker jobs at the expense of a more
complicated deployment.

The previous docker-compose used docker volumes for the database and
assetstore.  Because of docker permission weirdness, docker volumes
don't work once the container is run with a specific user, but running
with a specific user is needed for a docker to start another docker
container, so the assetstore and database are now stored in
subdirectories in the devops/annotation_tracker directory.  Migrating a
previous deployment would involve copying these files from the docker
volume to these subdirectories and ensuring that all files are owned by
the current user.