Add some specific annotation activities.
When the annotation activity occurs, it can be "delete", "edit" (new or
existing annotation is now being edited), "stopedit" (no annotation is
currently being edited), or "update" (some element has changed).  The id
of the current annotation is logged (but not the specific element that
changed).  The annotation version is logged, so if annotation history is
on, you can load the specific instance that was in use at that time.